requirements.txt
streamlit
google-genai
mcp
asyncio
fpdf
google-cloud-storage
requests
slack-sdk